[refactor] 자동차 경주, 출력 클래스 분리 : RacingCarGameRule 클래스 추가
